+++ R/process_pdf_county_testing.R
@@ -3,8 +3,10 @@ process_county_testing <- function(page_text, timestamp) {
     process_county_testing_v1(page_text, timestamp)
   } else if (timestamp >= "2020-03-20 17:40:00 EDT" && timestamp < "2020-03-21 17:31:00 EDT") {
     process_county_testing_v2(page_text, timestamp)
-  } else if (timestamp >= "2020-03-21 17:31:00 EDT") {
+  } else if (timestamp >= "2020-03-21 17:31:00 EDT" && timestamp < "2020-03-25 10:07:00 EDT") {
     process_county_testing_v3(page_text, timestamp)
+  } else if (timestamp >= "2020-03-25 10:07:00 EDT") {
+    process_county_testing_v4(page_text, timestamp)
   }
 }
 
@@ -50,6 +52,28 @@ process_county_testing_v3 <- function(page_text, timestamp) {
     ) %>% 
     drop_empty() %>%
     select(-contains("percent")) %>% 
+    replace_na(list(inconclusive = 0)) %>% 
+    add_total(from = c("negative", "positive", "inconclusive")) %>%  
+    filter(county != "Total") %>% 
+    add_timestamp(timestamp)
+}
+
+process_county_testing_v4 <- function(page_text, timestamp) {
+  pages_testing <- page_text %>% 
+    map_lgl(str_detect, pattern = "All persons with test") %>% 
+    which()
+  
+  page_text[pages_testing] %>%
+    # remove last two columns (total and percent)
+    str_replace_all("[\\d,]+\\s+\\d+%\n", "\n") %>% 
+    read_table_pages(
+      col_names = c("county", "pending", "inconclusive", "negative", "positive"),
+      pattern_tab = "\\s{3,}"
+    ) %>% 
+    filter(county != "Total") %>% 
+    drop_empty() %>%
+    select(-contains("percent")) %>% 
+    mutate_at(vars(pending, inconclusive, negative, positive), as.numeric) %>% 
     replace_na(list(inconclusive = 0)) %>% 
     add_total(from = c("negative", "positive", "inconclusive")) %>%  
     filter(county != "Total") %>%